% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{estimate_variance_w_args}
\alias{estimate_variance_w_args}
\title{estimate_variance_w_args}
\usage{
estimate_variance_w_args(pars, counts, total_cells, optim_clusters)
}
\arguments{
\item{pars}{A numeric vector of length equal to no. of samples, representing their corresponding normalization factors}

\item{counts}{A cell count matrix with number of columns representing the number of samples and number of rows representing the number of cell-types or clusters}

\item{total_cells}{An integer vector representing the total number of cells per sample}

\item{optim_clusters}{An integer vector representing the indices of the clusters over which the total variance is to be calculated}
}
\value{
A numeric value representing the sum of variances of log odds of cluster membership across all clusters specified in optim_clusters
}
\description{
Total (across a specified set of clusters) of the sample variance of the log odds of clusters membership across all samples within each of those clusters
}
