% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{ineqfun_data}
\alias{ineqfun_data}
\title{ineqfun_data}
\usage{
ineqfun_data(pars, total_cells)
}
\arguments{
\item{pars}{A numeric vector of length equal to no. of samples, representing their corresponding normalization factors}

\item{total_cells}{An integer vector representing the total number of cells per sample}
}
\value{
A numeric vector representing the normalized number of total cells per sample
}
\description{
Normalized number of total cells per sample
}
