% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{logodds_optimized_normFactors}
\alias{logodds_optimized_normFactors}
\title{logodds_optimized_normFactors}
\usage{
logodds_optimized_normFactors(cellcomp_se, verbose = TRUE)
}
\arguments{
\item{cellcomp_se}{a SummarizedExperiment object that includes
\itemize{
\item counts: an assay named counts with the number of cells from each sample in each of the clusters. The counts is a data frame with rownames set to the cluster names (without spaces or special characters) and column names representing the sample names
\item data for the association analyses assigned to colData as a data frame with row names the same as the column names of the counts data frame
\item Five parameters assigned to the metadata slot
\itemize{
\item modelFormula representing the model to be tested using the variable in the colData
\item coef_of_interest_index representing the index of the coefficient of interest to be tested in the model
\item reference_levels_of_variables is a list of vectors.Each vector will have two elements - the name of a categorical variable (a column name in colData) and the reference level to set this variable to
\item random_seed integer random seed
\item unchanged_cluster_indices cluster indices of the clusters known not to change, set to NULL if not known
}
}}

\item{verbose}{Logical; if TRUE (default), prints progress messages}
}
\value{
A list with two elements
\itemize{
\item results: data frame of results of differential analyses with 4 columns - cluster_id, comparison, estimates (log odds ratio) and estimate_significance (pvalue)
\item optim_factor: a data frame with columns sampleid and optim.norm.factor
}
}
\description{
The primary LODopt method
}
\examples{
# Create example data
set.seed(123)
# No of clusters in single cell dataset
K <- 25
nsamp <- 30
alpha <- 10^runif(K, min = log10(0.5), max = log10(10))
p <- dirmult::rdirichlet(alpha = alpha) |> sort()
p <- p[p > 0.001]
p <- p / sum(p)
size <- rep(10, length(p))
change_mean <- rep(1, length(p))
## clusters 1, 3, 8 and 15 are changed. Clusters 1 and 8 reduce in abundance while clusters 3 and 15 increase
change_mean[c(1, 3, 8, 15)] <- c(0.2, 2, 0.2, 2)
depth <- 1e9
# Simulate counts
counts_res <- simulate_cellCounts_fromTissue(
  props = p,
  nsamp = nsamp,
  depth = depth,
  size = size,
  change_mean = change_mean
)
counts <- counts_res$counts

pheno_data <- data.frame(
  sampleID = paste0("S", 1:30),
  groupid = c(rep("group0", 15), rep("group1", 15))
)
require(magrittr)
pheno_data \%<>\% tibble::column_to_rownames("sampleID")
require(SummarizedExperiment)
model_formula <- "groupid"
cellcomp_se <- SummarizedExperiment(
  assays = list(counts = counts),
  colData = pheno_data,
  metadata = list(
    modelFormula = model_formula,
    coef_of_interest_index = 2,
    reference_levels_of_variables = list(c("groupid", "group0")),
    random_seed = 123456,
    unchanged_cluster_indices = NULL
  )
)
cellcomp_res <- logodds_optimized_normFactors(cellcomp_se)
print(cellcomp_res$results)

}
